+++ Changelog.md
@@ -2,6 +2,13 @@
 
 All Merges to this project will be documented in this file.
 
+## [Version 1.0.1] - 2025-05-25
+
+## Added
+- Added the tool to encrypt the secret file.
+- Added functionality to decrypt the secret file with password.
+
+
 ## [Version 1.0.0] - 2025-03-14
 
 ### Added